Completing the basic login UI
