+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'login.dart';
 
 void main() => runApp(new MyApp());
 
@@ -9,96 +10,6 @@ class MyApp extends StatelessWidget {
       home: new LoginPage(),
       theme: new ThemeData(
         primarySwatch: Colors.green
-      ),
-    );
-  }
-}
-
-class LoginPage extends StatefulWidget {
-  @override
-  State createState() => new LoginPageState();
-}
-
-class LoginPageState extends State<LoginPage> with SingleTickerProviderStateMixin {
-  
-  AnimationController _iconAnimationController;
-  Animation<double> _iconAnimation;
-
-  @override
-  void initState() {
-    super.initState();
-    
-    _iconAnimationController = new AnimationController(
-      vsync: this,
-      duration: new Duration(milliseconds: 500)
-    );
-
-    _iconAnimation = new CurvedAnimation(
-      parent: _iconAnimationController,
-      curve: Curves.bounceOut
-    );
-
-    _iconAnimation.addListener(() => this.setState((){}));
-    _iconAnimationController.forward();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return new Scaffold(
-      backgroundColor: Colors.greenAccent,
-      body: new Stack(
-        fit: StackFit.expand,
-        children: <Widget>[
-          new Image(
-            image: new AssetImage("assets/login_background.jpg"),
-            fit: BoxFit.cover,
-            color: Colors.black54,
-            colorBlendMode: BlendMode.darken,
-          ),
-          new Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget> [
-              new FlutterLogo(
-                size: _iconAnimation.value * 100
-              ),
-              new Form(
-                child: new Theme (
-                  data: new ThemeData (
-                    brightness: Brightness.dark,
-                    primarySwatch: Colors.green,
-                    inputDecorationTheme: new InputDecorationTheme(
-                      labelStyle: new TextStyle (
-                        color: Colors.green,
-                        fontSize: 20
-                      )
-                    )
-                  ),
-                  child: Container(
-                    padding: const EdgeInsets.all(40.0),
-                    child: new Column(
-                      crossAxisAlignment: CrossAxisAlignment.center,
-                      children: <Widget> [
-                        new TextFormField(
-                          decoration: new InputDecoration(
-                            hintText: "Email",
-                          ),
-                          keyboardType: TextInputType.emailAddress,
-                        ),
-                        new TextFormField(
-                          decoration: new InputDecoration(
-                            hintText: "Password",
-                          ),
-                          keyboardType: TextInputType.text,
-                          obscureText: true,
-                        ),
-                      ] 
-                    ),
-                  ),
-                ),
-              )
-            ]
-          )
-        ],
       ),
     );
   }
